Refactor solver dependency and goal types.
This commit should have no effect on behavior.  It splits the Dep type into two
new types, PackageDep and Constraint.  A PackageDep is a dependency that
introduces a goal, i.e., it is something that the solver can choose to install.
A Constraint is a dependency that doesn't introduce a goal, such as a pkg-config
or language extension dependency.  The solver must check that the constraint
holds, but it cannot do anything to make it hold.

Per-component solving will introduce more dependencies that behave like
constraints, so it is important to no longer treat them as special cases.

I modified the FlaggedDep type to distinguish between the two types of
dependencies.  The main changes are in Dependency.hs and Builder.hs.  I also
made some smaller changes:

- Add a type, OpenDependencyGoal, that is used instead of OpenGoal in places
  where a dependency must introduce a goal.
- Combine FlaggedDep's flag and stanza constructors, since they are handled
  similarly.
- Remove OpenGoal and FlaggedDep's 'comp' type variables.  OpenDependencyGoal
  replaces the use of OpenGoal and FlaggedDep with 'comp' equal to ().  The only
  remaining uses of OpenGoal and FlaggedDep use Component.
- Move OpenGoal to Builder.hs.
- Rename FlagInfo to FInfoMap.