Solver: Check whether packages are buildable in the current environment.
This commit handles the most common case of issue #5325 by checking that each
package is buildable in the current environment.  The solver replaces
non-buildable packages with failure nodes.  It allows two types of buildable
packages:

Buildable non-test package:
The package has at least one buildable library or executable.

Buildable test package:
The package has no libraries or executables, but it does have at least one
buildable test or benchmark.

The buildable check can give false-positives, because it ignores all flags, and
it doesn't check whether the intra-package dependencies of a component are
buildable.  The check is also incomplete because it is performed before any
flags are assigned.  It is possible for the solver to later choose a value for a
flag that makes the package unbuildable.